+++ env.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+source ./env/bin/activate
