+++ test.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+source ./env/bin/activate
+python -m pytest
